+++ CHANGELOG.md
@@ -30,6 +30,7 @@
       },
     });
     ```
+
 - Add `strictTraceContinuation` and `orgId` options for trace continuation validation ([#1166](https://github.com/getsentry/sentry-capacitor/pull/1166))
 
 ### Dependencies
@@ -40,9 +41,9 @@
 - Bump JavaScript Sibling SDKs from v9.42.0 to v10.52.0 ([#1244](https://github.com/getsentry/sentry-capacitor/pull/1244))
   - [changelog](https://github.com/getsentry/sentry-javascript/blob/10.52.0/CHANGELOG.md)
   - [diff](https://github.com/getsentry/sentry-javascript/compare/10.42.0...10.52.0)
-- Bump Cocoa SDK from v9.8.0 to v9.13.0 ([#1248](https://github.com/getsentry/sentry-capacitor/pull/1248))
-  - [changelog](https://github.com/getsentry/sentry-cocoa/blob/main/CHANGELOG.md#9130)
-  - [diff](https://github.com/getsentry/sentry-cocoa/compare/9.8.0...9.13.0)
+- Bump Cocoa SDK from v9.8.0 to v9.16.0 ([#1248](https://github.com/getsentry/sentry-capacitor/pull/1248))
+  - [changelog](https://github.com/getsentry/sentry-cocoa/blob/main/CHANGELOG.md#9160)
+  - [diff](https://github.com/getsentry/sentry-cocoa/compare/9.8.0...9.16.0)
 
 ## 4.0.0